Refactor SourceAssets implementation
Pure refactor Commit - there should be no behaviour change.
Split the data structures and analysis away from the UI, similar to how ContentSummary is written.
Add a namespace to ContentSummmary code.
