+++ com.unity.build-report-inspector/Editor/SourceAssets.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 66c7453628aa2b543b391dc681699ff9
